move extension traits to ext.rs
Amp-Thread-ID: https://ampcode.com/threads/T-019ca92f-974d-72be-9042-3692bd735712
